chore(TimePicker): remove unused TimeOption compone and select dependencies
