+++ packages/react-core/src/components/Select/SelectOption.tsx
@@ -3,9 +3,10 @@ import styles from '@patternfly/react-styles/css/components/Select/select';
 import checkStyles from '@patternfly/react-styles/css/components/Check/check';
 import { css } from '@patternfly/react-styles';
 import CheckIcon from '@patternfly/react-icons/dist/esm/icons/check-icon';
-import { SelectConsumer, SelectVariant, KeyTypes } from './selectConstants';
+import { SelectConsumer, SelectVariant } from './selectConstants';
 import StarIcon from '@patternfly/react-icons/dist/esm/icons/star-icon';
 import { getUniqueId } from '../../helpers/util';
+import { KeyTypes } from '../../helpers/constants';
 
 export interface SelectOptionObject {
   /** Function returns a string to represent the select option object */
