+++ packages/react-core/src/components/TimePicker/index.ts
@@ -1,2 +1,1 @@
 export * from './TimePicker';
-export * from './TimeOption';
