+++ packages/react-core/src/helpers/constants.ts
@@ -10,3 +10,14 @@ export enum ValidatedOptions {
   warning = 'warning',
   default = 'default'
 }
+
+export const KeyTypes = {
+  Tab: 'Tab',
+  Space: ' ',
+  Escape: 'Escape',
+  Enter: 'Enter',
+  ArrowUp: 'ArrowUp',
+  ArrowDown: 'ArrowDown',
+  ArrowLeft: 'ArrowLeft',
+  ArrowRight: 'ArrowRight'
+};
